Allow user to specify output format
The default `cli-table` implementation does not look great when using a non-fixedwidth font (as is the case in many IRC clients and in Slack). This PR introduces a format abstraction that switches based on a `HUBOT_UBER_OUTPUT_FORMAT`. It will default to unformatted (splits the values with a `/` when listing them).